debug/keraspipeline/requirements.txt
absl-py==0.5.0
astor==0.7.1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.2
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Keras==2.2.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
mistune==0.8.3
numpy==1.15.2
Pillow==5.3.0
protobuf==3.6.1
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
scipy==1.1.0
six==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1